requirements.txt
Flask
tensorflow
opencv-python
Pillow
numpy
flask_cors